chore: added granular otel apps
