+++ example-apps/otel/serverless-otlp-format-grpc/README.md
@@ -1,11 +1,10 @@
-# Otel playground
+# Serverless OpenTelemetry OTLP traces and metrics
 
 - Choose an [OTLP acceptor endpoint](https://www.ibm.com/docs/en/instana-observability/current?topic=instana-backend#endpoints-of-the-instana-backend-otlp-acceptor).
 - Read the x-instana-key from the agent installation docs of the target SaaS instance.
-- You can also just start and play without adding the endpoint and key.
 
 ```sh
-OTEL_EXPORTER_OTLP_HEADERS="x-instana-key=xxxx" OTEL_EXPORTER_OTLP_ENDPOINT=endpoint OTEL_LOG_LEVEL=debug OTEL_SERVICE_NAME=my-js-service node server.js
+INSTANA_KEY=xxxx OTEL_EXPORTER_OTLP_ENDPOINT=otlp-green-saas.instana.io:4317 OTEL_LOG_LEVEL=debug OTEL_SERVICE_NAME=node-otlp-grpc npm start
 
 node kafka-consumer.js
 ```
