+++ example-apps/otel/serverless-otlp-format-grpc/kafka-consumer.js
